+++ scripts/world_location.gd
@@ -32,5 +32,5 @@ func set_ui(_ui : UI):
 func get_player_car():
 	pass
 
-func set_player_car(player_car_id : int):
+func set_player_car(player_car_id : String):
 	pass